ACTUALLY removed tests for removed endpoints
